# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Alaa API",
        "description": "An API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/roles/{role_name}": {
            "get": {
                "tags": [
                    "Role Management"
                ],
                "summary": "Get role by name",
                "operationId": "f9d964c88b482db108e45bc83f320186",
                "parameters": [
                    {
                        "name": "role_name",
                        "in": "path",
                        "description": "The name of the role to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "role_name": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Role not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Role not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/roles": {
            "get": {
                "tags": [
                    "Role Management"
                ],
                "summary": "Get all roles",
                "operationId": "1811fddab2cec2e9cd45d508ee99b560",
                "responses": {
                    "200": {
                        "description": "List of all roles"
                    },
                    "404": {
                        "description": "No roles found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No roles found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/roles/role/{role_id}": {
            "get": {
                "tags": [
                    "Role Management"
                ],
                "summary": "Get role by ID",
                "operationId": "a7411ff8813f469d3368a21bf145ed48",
                "parameters": [
                    {
                        "name": "role_id",
                        "in": "path",
                        "description": "The ID of the role to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "role_name": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Role not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Role not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/roles/create/role": {
            "post": {
                "tags": [
                    "Role Management"
                ],
                "summary": "Create a new role",
                "operationId": "a29ccf3471d01171432679dd93515514",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "role_name": {
                                        "type": "string",
                                        "example": "Admin"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Administrator role with full permissions"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Role created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Role created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Role creation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Role could not be created"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/roles/update/{role_id}": {
            "put": {
                "tags": [
                    "Role Management"
                ],
                "summary": "Update an existing role",
                "operationId": "85c50f47c2c20104cc8c17c5d8dc82f7",
                "parameters": [
                    {
                        "name": "role_id",
                        "in": "path",
                        "description": "The ID of the role to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "role_name": {
                                        "type": "string",
                                        "example": "Admin"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Administrator role with full permissions"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Role updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Role updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Role not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Role not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Role update failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Role could not be updated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/roles/delete/{role_id}": {
            "delete": {
                "tags": [
                    "Role Management"
                ],
                "summary": "Delete a role",
                "operationId": "c66414e2bb4da3c2bc7734b63b2ba71f",
                "parameters": [
                    {
                        "name": "role_id",
                        "in": "path",
                        "description": "The ID of the role to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Role deleted successfully"
                    },
                    "404": {
                        "description": "Role not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Role not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}